requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.7.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.1
pytz==2024.1
